part_3_agents/chapter_13_agent_memory/requirements.txt
# requirements.txt
python-dotenv==1.2.1
openai==2.9.0
chromadb==1.3.5
numpy==2.3.5
langchain==1.1.3
langchain-openai==1.1.1
langchain-community==0.4.1
langchain-classic==1.0.0